backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.4",
    "huggingface-hub>=1.2.3",
    "httpx>=0.27.0",
    "pyngrok>=7.5.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
